Match 7.79% of `ftCo_0A01` (81.08 KB) (#1417)

+++ src/melee/pl/player.c
@@ -791,9 +791,9 @@ void Player_SetPlayerId(int slot, s8 player_id)
     player->player_id = player_id;
 }
 
-s8 Player_GetCpuLevel(int slot)
+int Player_GetCpuLevel(int slot)
 {
-    s8 cpu_level;
+    int cpu_level;
     StaticPlayer* player;
     Player_CheckSlot(slot);
     player = &player_slots[slot];
@@ -801,7 +801,7 @@ s8 Player_GetCpuLevel(int slot)
     return cpu_level;
 }
 
-void Player_SetPlayerAndEntityCpuLevel(int slot, s8 cpu_level)
+void Player_SetPlayerAndEntityCpuLevel(int slot, int cpu_level)
 {
     StaticPlayer* player;
     int i;
@@ -817,9 +817,9 @@ void Player_SetPlayerAndEntityCpuLevel(int slot, s8 cpu_level)
     }
 }
 
-s8 Player_GetCpuType(int slot)
+int Player_GetCpuType(int slot)
 {
-    s8 cpu_type;
+    int cpu_type;
     StaticPlayer* player;
     Player_CheckSlot(slot);
     player = &player_slots[slot];
@@ -827,7 +827,7 @@ s8 Player_GetCpuType(int slot)
     return cpu_type;
 }
 
-void Player_SetPlayerAndEntityCpuType(int slot, s8 cpu_type)
+void Player_SetPlayerAndEntityCpuType(int slot, int cpu_type)
 {
     StaticPlayer* player;
     int i;
