+++ src/melee/pl/player.h
@@ -48,8 +48,8 @@ typedef struct _StaticPlayer {
     /*0x46*/ s8 controller_index;
     /*0x47*/ u8 team; /// 00 = red, 01 = blue, 02 = green
     /*0x48*/ s8 player_id;
-    /*0x49*/ s8 cpu_level;
-    /*0x4A*/ s8 cpu_type;
+    /*0x49*/ u8 cpu_level;
+    /*0x4A*/ u8 cpu_type;
     /*0x4B*/ s8 handicap;
 
     /*0x4C*/ s8 unk4C;
@@ -194,10 +194,10 @@ s8 Player_GetTeam(int slot);
 void Player_SetTeam(int slot, s8 unk_color);
 s8 Player_GetPlayerId(int slot);
 void Player_SetPlayerId(int slot, s8 player_id);
-s8 Player_GetCpuLevel(int slot);
-void Player_SetPlayerAndEntityCpuLevel(int slot, s8 cpu_level);
-s8 Player_GetCpuType(int slot);
-void Player_SetPlayerAndEntityCpuType(int slot, s8 cpu_type);
+int Player_GetCpuLevel(int slot);
+void Player_SetPlayerAndEntityCpuLevel(int slot, int cpu_level);
+int Player_GetCpuType(int slot);
+void Player_SetPlayerAndEntityCpuType(int slot, int cpu_type);
 int Player_GetHandicap(int slot);
 void Player_SetHandicap(int slot, s8 handicap);
 f32 Player_GetUnk50(int slot);
